+++ src/main/kotlin/yandex/cloud/toolkit/api/service/impl/CloudRepositoryImpl.kt
@@ -15,6 +15,9 @@ import yandex.cloud.api.logging.v1.LogGroupServiceOuterClass.ListLogGroupsReques
 import yandex.cloud.api.logging.v1.LogReadingServiceGrpc
 import yandex.cloud.api.logging.v1.LogReadingServiceOuterClass.Criteria
 import yandex.cloud.api.logging.v1.LogReadingServiceOuterClass.ReadRequest
+import yandex.cloud.api.lockbox.v1.PayloadOuterClass
+import yandex.cloud.api.lockbox.v1.PayloadServiceGrpc
+import yandex.cloud.api.lockbox.v1.PayloadServiceOuterClass
 import yandex.cloud.api.operation.OperationOuterClass
 import yandex.cloud.api.operation.OperationServiceGrpc
 import yandex.cloud.api.operation.OperationServiceOuterClass
@@ -268,6 +271,17 @@ class CloudRepositoryImpl : CloudRepository {
                         FunctionOuterClass.Connectivity.newBuilder().setNetworkId(spec.networkId).build()
                     }
                 }
+
+                for (secret in spec.secrets) {
+                    addSecrets(
+                        FunctionOuterClass.Secret.newBuilder()
+                            .setId(secret.id)
+                            .setVersionId(secret.versionId)
+                            .setKey(secret.key)
+                            .setEnvironmentVariable(secret.envVariable)
+                            .build()
+                    )
+                }
             }.build()
 
             authData().functionService.createVersion(request)
@@ -306,6 +320,17 @@ class CloudRepositoryImpl : CloudRepository {
         }
 
         return CloudOperation("Remove scaling policy '$tag'", operation)
+    }
+
+    override fun getLockboxSecret(
+        authData: CloudAuthData,
+        secretId: String,
+        versionId: String?
+    ): PayloadOuterClass.Payload {
+        val request = PayloadServiceOuterClass.GetPayloadRequest.newBuilder()
+        request.secretId = secretId
+        if (versionId != null) request.versionId = versionId
+        return authData().payloadService.get(request.build())
     }
 
     override fun getOperation(authData: CloudAuthData, operationId: String): Maybe<OperationOuterClass.Operation> {
@@ -798,6 +823,12 @@ class CloudRepositoryImpl : CloudRepository {
             createService(
                 BucketServiceGrpc.BucketServiceBlockingStub::class.java,
                 BucketServiceGrpc::newBlockingStub
+            )
+        }
+
+        val payloadService by lazy {
+            createService(
+                PayloadServiceGrpc.PayloadServiceBlockingStub::class.java, PayloadServiceGrpc::newBlockingStub
             )
         }
     }