+++ src/main/kotlin/yandex/cloud/toolkit/configuration/function/deploy/DeployFunctionConfiguration.kt
@@ -21,7 +21,7 @@ import yandex.cloud.toolkit.process.RunContentController
 import yandex.cloud.toolkit.util.logger
 
 class DeployFunctionConfiguration(name: String?, factory: ConfigurationFactory, project: Project) :
-    RunConfigurationMinimalBase<FunctionDeploySpec>(name, factory, project) {
+    RunConfigurationMinimalBase<FunctionDeploySpec>(name, factory, project), WithoutOwnBeforeRunSteps {
 
     override fun clone() = DeployFunctionConfiguration(name, factory, project).apply {
         state.copyFrom(this@DeployFunctionConfiguration.state)
@@ -98,6 +98,7 @@ class FunctionDeploySpec : BaseState() {
     var objectStorageBucket by string()
     var objectStorageObject by string()
     var updateObjectStorage by property(false)
+    var secrets by list<FunctionSecret>()
 
     fun hasConnectivity(): Boolean = if (useSubnets) subnets.isNotEmpty() else !networkId.isNullOrEmpty()
 
@@ -126,6 +127,28 @@ class FunctionDeploySpec : BaseState() {
                 useSubnets = networkId.isNullOrEmpty()
                 subnets = template.connectivity.subnetIdList
             }
+            secrets = template.secretsList.map(FunctionSecret.Companion::from).toMutableList()
+        }
+    }
+}
+
+class FunctionSecret : BaseState() {
+    var id by string()
+    var versionId by string()
+    var key by string()
+    var envVariable by string()
+
+    override fun toString(): String {
+        return "$envVariable = $id -> $versionId -> $key"
+    }
+
+    companion object {
+
+        fun from(secret: FunctionOuterClass.Secret) = FunctionSecret().apply {
+            id = secret.id
+            versionId = secret.versionId
+            key = secret.key
+            envVariable = secret.environmentVariable
         }
     }
 }